Refactored InstallReferrerClientWrapper
